pyproject.toml
[project]
name = "pixelpet"
version = "0.1.0"
description = "Pixel desktop pet for macOS (PyObjC/AppKit)"
requires-python = ">=3.11,<3.12"
readme = "README.md"
authors = [{ name = "Tori" }]
dependencies = [
  "pyobjc>=10.1",
  "pillow>=10.0",
  "fastapi>=0.110",
  "uvicorn>=0.29",
  "httpx>=0.27",
]

[project.scripts]
pixelpet = "app.main:main"
pixelpet-bridge = "app.openclaw_bridge:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
package = true
